+++ src/test/java/core/di/factory/ConfigurationBeanScannerTest.java
@@ -1,4 +1,4 @@
-package next.config;
+package core.di.factory;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,6 +6,7 @@ import javax.sql.DataSource;
 
 import org.junit.jupiter.api.Test;
 
+import core.di.ConfigurationBeanScanner;
 import core.di.factory.BeanFactory;
 import core.di.factory.example.ExampleConfig;
